requirements.txt
requests==2.31.0
cloudscraper==1.2.71
beautifulsoup4==4.12.2
yfinance==0.2.18
finnhub-python==2.4.19
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
ta==0.10.2
pandas-ta==0.3.14b0
python-dotenv==1.0.0
pydantic==2.4.2
aiohttp==3.8.5
asyncio-throttle==1.0.2
tkinter-tooltip==2.1.0